added register controller
