+++ controllers/signin.js
@@ -1,0 +1,26 @@
+const handleSignin = (req, res, db, bcrypt) => {
+  const { email, password } = req.body;
+  db.select("email", "hash")
+    .from("login")
+    .where("email", "=", email)
+    .then((data) => {
+      const isValid = bcrypt.compareSync(password, data[0].hash);
+      if (isValid) {
+        return db
+          .select("*")
+          .from("users")
+          .where("email", "=", email)
+          .then((user) => {
+            res.json(user[0]);
+          })
+          .catch((error) => res.status(400).json("unable to get user"));
+      } else {
+        res.status(400).json("Incorrect credentials");
+      }
+    })
+    .catch((error) => res.status(400).json("wrong credentials"));
+};
+
+module.exports = {
+  handleSignin,
+};
